Refactored how characters are placed into scenes
Now characters are instanced by GodotAdventureQuest and it checks if the
character is in the current room to add it to the scene.
Player character is added by default to any Room. An inspector variable
was created to avoid this: Room.has_player.
Improved how characters move. Now they don't listen to a signal to check
if it has to move, instead, Character.walk(...) can be called and yield
directly.