Created Globals script
Rooms are now loaded in GodotAdventureQuest.gd as a Resource with the ID
to load the room and the path to its scene.
Room scripts and scenes renamed as Room(name of the room).gd and
Room(name of the room).tscn.
RoomForest can delete the bucket if the player already picked it or used
it in the well (this is a test for the Globals script).