Cutscenes can be skipped and a transition is shown.
Added new type of transition: pass down.
The created transition is used when a cutscene is skipped.
Character walk and say methods can be skipped if they are part of a
cutscene.
Graphic Interface DisplayBox can be skipped during cutscene.
The events that are part of a cutscene trigger on skip. This makes it
possible to make changes in the room or the game even when the cutscene
is skipped.
--- Ready to work on Toolbar ---