Created Dock plugin to ease the creation of Rooms.
When creating a Room with the new Dock, three files are created: a .tscn
with a .gd and the .tres that is used by GodotAdventureQuest.tscn to
register the Rooms that will be available in the game.
Once the files for the Room are created, the .tres is added to
GodotAdventureQuest's rooms property.
Once the process has successfully finished, the created scene for the
Room is opened in the Editor.
Created a new type of resource: GAQCharacter. This is used by
GodotAdventureQuest to have the list of available characters.
Now characters are instanced only when a scene requires them.
Soon the name of the plugin will change... because of course it has to
be in Colombian.
--- Ready to work on the plugins to ease the creation of Prop, Hotspot
and Region. ---