Plugin can now create dialog trees.
The created dialog tree is selected in the tree and "opened" for
edition.
Updated the guide text for all the popups to make clear that the naming
convention is PascalCase in English.
Updated the guide text for all popups to include better examples of
filenames in each type.
(!) All dialog trees are now loaded in GodotAdventureQuest.tscn.
DialogTreen.tscn has dissappeared.
Now only the IDialogTree.gd is used for the D autoload.
Updated the file structure for the dialog trees.