Plugin can now create Room Props.
The created node may or may not have interaction, creating a script in the first case.
! Added tabs to the plugin dock. The Main tab contains lists of Rooms,
Characters, Inventory items and dialog trees (soon AudioCues); the Room
tab contains lists of Props, Hotspots (and soon Regions) of a selected
Room in the editor.
Created temporary icons for props and hotspots.
Minor visual enhancenments.
\(9 9 D)/