Popochiu's object folders can be recursiverly deleted
This can be useful for deleting the objects created by the plugin plus
the imported files associated to them: images, audio files, etc.
Created a generic deletion confirmation dialog in PopochiuDock to be
used when deleting an Popochiu object.
Created first pair of DynamicFont to use as the main font for the
plugin.
Asking the developer if a foldar must be deleted is now part of the
confirmation dialog.
--- Ready to work on making the plugin to work standalone ---