Making the plugin to work on empty projects.
Moved the GraphicInterface and TransitionLayer scenes to the addon
folder.
Moved all PNG files from assets/* to their respective folders where they
are used.
All objects in the filesystem that are not in Popochiu core can be added
to it using the dock.
Improved how transition animations are setup.
Working on creating the folders and files required to the plugin to work
in an empty project.
Fixed some dependency issues in Clickable object.