Added restart popup on plugin enabling.
Restarting the engine will ensure everything works de chupelupe.
Updated the scripts of main objects creation to use a common method in
PopochiuDock to add their respective resource to their corresponding
Array in Popochiu.
Fixed a bug on creating main objects for the first time in Popochiu.
Fixed a bug on adding objects to Popochiu using the PopochiuObjectRow
Add to core button.
Fixed a bug of duplicated lines in main objects' lists when creating
them for the first time.
=== Ready to work assigning the GraphicInterface and TransitionLayer
during install ===