Audio tab rows can play audio streams in AudioCues
The play and stop buttons in the AudioCue row are using the play/pause
and stop icons from the Godot editor.
Changes in the AudioCure .tres file can be heard immediatly in the
AudioCue row.
Ready to work on opening the AudioCue in the Inspector when clicking
play or when clicking the button that does that in PopochiuObjectRow.
Added AudioStreamPlayers to the Audio tab in the PopochiuDock.
Created icons to assign an audio file to an Audio manager category:
Music, SFX, Voice and Graphic interface.