Audio files without category can be played and deleted
Deletion in unassigned files deletes the file from the system.
Playing unassigned audio files ignores their loop property.
PopochiuObjectGroup now shows the number of items in the group if it is
greater than 2.
The plugin listens changes in the project sources to append new audio
files to the Unassigned category when a new audio file is found.
AudioCues can be deleted both from the AudioManager and the file system.
AudioCue .tres files are open when the row in the inspector is clicked,
played or stopped.
=== Ready to work on auto assigning audio files based on their prefix
===