PopochiuGroupButton is now PopochiuGroup.
This reduces the complexity related to the list of items related to the
group and the button used to create new items for it.
Removed deprecated code from PopochiuDock thanks to this refactor.
Now the open/close icon used in the group is the same used by Godot for
grouped controls.
=== Ready to work on moving the scripts related to Room and Audio tabs
to their own files ===