Room tab refactor.
Now Props are created inside a folder with the name of the prop.
Great improvement in the flow that readed the objects in the Room to put
them in the dock.
Now rows of Room objects in the dock are PopochiuObjectRow, instead of a
Label.
Fix: The game crashed when receiving an empty array of instructions in
E.run().
=== All ready to prepare an up to date folder of the plugin ===