Folders deletion working.
This fixes a bug that was causing some subfolders to not be deleted
during the process, which broke the deletion of rooms.
Developers are notified of deleting audio file when deleting a folder.
Fix: Popochiu reacts better to trying to move to a not existent room. It
stays at the current room when so happens.
=== Ready to work on reassigning the main scene when deleting a Room set
as the main scene ===