* Removed animation dependencies in PopochiuCharacter Now this Object only has a Sprite. Animations will be a responsability of developers. Inspector allows to define how to flip the character Sprite when moving. By default is set to None. Made a minor improvement in ICharacter so developers are not responsible of disabling the GI when calling `character_say` or `player_say`. Replaced some code comments with English version.
* Inventory items can be added to inventory on start
This can be done using the context menu in each inventory item row in the Main tab.
Re-imported all pixel art images with the 2D Pixel preset.

* Updated how Prop's z_index is updated on Character movement.
Added more Props and a Character to the project for testing.
FIX: Replaced Array.remove(...) by Array.erase(...) while deleting inventory items from E.items_in_start. Don't know how was it working anyway.

* Props, Hotspots and Regions can be deleted from dock.
Now Hotspots and Regions are created within folders to make easier their deletion.
Object rows in Main and Room tabs now show the path to the object.

* Room objects can be deleted from Room tab.
FIX: The search audio files button in Audio tab was not working.

* Added Popochiu prefix to some base Objects.
Removed references to Popochiu in some plugin's nodes so the Output does
not show all those annoying bugs when using the plugin for the first
time.
Removed the default CollisionPolygon2D from Character, Hotspot and Prop
so the nodes doesn't have to be set as Editable when adding them to a
scene.

* Removed class_name for some scripts
It was replaced by `preload()`s.
Clickable now has a reference to the room to which it belongs.
Created Constants.gd to store all plugin related constants there.