Prop z_index updating is now based in the baseline.
When a Prop has to change its z_index, other props in front of it change their z_index too. This is based on the order of the Props determined by its baseline.
C singleton now converts the position to move a character to into global based on the current room (E.current_room).
Progress in translating code comments to English.
Deleted old files.