Added camera related fuctions to Popochiu (E).
Developers can now shake, zoom and offset the camera: camera_shake(), camera_zoom(), camera_offset().
Moved show_inline_dialog to IDialog.
Camera is not linked to C.player but C.camera_owner, which can be changed with C.change_camera_owner(). By default C.player is set as the camera owner.
Updated how the game's viewport width, height, half_widht and half_height are get. Now this are not onready vars, but public vars with get methods.
Added enum for transition types to TransitionLayer.gd.
Added room_exists method to E. This allows to check if a room existis with a script_name.
Added play_transition method to E. This allows to play a transition of one of the types defined in TransitionLayer.gd.
Updated the interaction with PropFileCabinet to show some of this additions in action.