Dev update (#5)
* Added camera related fuctions to Popochiu (E).

Developers can now shake, zoom and offset the camera: camera_shake(), camera_zoom(), camera_offset().
Moved show_inline_dialog to IDialog.
Camera is not linked to C.player but C.camera_owner, which can be changed with C.change_camera_owner(). By default C.player is set as the camera owner.
Updated how the game's viewport width, height, half_widht and half_height are get. Now this are not onready vars, but public vars with get methods.
Added enum for transition types to TransitionLayer.gd.
Added room_exists method to E. This allows to check if a room existis with a script_name.
Added play_transition method to E. This allows to play a transition of one of the types defined in TransitionLayer.gd.
Updated the interaction with PropFileCabinet to show some of this additions in action.

* FIX: E.game_width was still in use in DialogText.gd.

Created class_name for TransitionLayer.gd.

* Inventory can have a limit defined as an export variable in E.

Renamed various methods, signals and variables names.
Moved signals and functions to show-hide the inventory from IGraphicInterface to IInventory.

* AudioManager refactorization.

Simplified the way play() works and added a couple of methods to modify AudioCues by thier script_name.

* Fix a bug with PopochiuAudioRow because a name change in semitone_to_pitch.

Added a sound effect and music to test the new play methods in A.

* Changed vars and params names for some functions in A.

FIX: When selecting a PopochiuRoom with Points the Room tab was not showing the points.
FIX: When selecting a Point in the Room tab the output was showing an error.

* Added Footer panel and button to open docs

* Translated comments to English.

Added some function comments based in those created for the Wiki.
FIX: PopochiuCharacter was calling A.play() with the old parameters.

* Deleted dev files