v1.8.4.a
==== v1.8.4.a ====
Added method to change the frame of a prop inside an E.run([]).
Updated templates to make easier to deleted commented paragraphs.

FIX: Typo in default on_item_used message.
FIX: Make the PC the character that is currently marked as the current player.

==== v1.8.4 ====
Now the Perimeter (NavigationPolygonInstance) is created by code.

FIX: Clicking on a row in Room tab was opening the script of the object. This was removed because there is a button inthe row to open the script.
FIX: Perimeter node in PopochiuWalkableArea scene was causing problems

==== v1.8.3 ====
Plugin version is shown in the footer of the dock.
Added export properties to PopochiuProp to change the number of `hframes` in its Sprite and the current frame.
Updated icons, labels and styles for "Baseline" and "Walk to point" buttons in the canvas.
Updated shoes color for walkable_area icon.

Thanks @StickGrinder for your work on PopochiuWalkableAreas
Added Walkable Areas list in Room tab.
Added basic PopochiuWalkableArea object and refactored short WA naming for consistency between UI and code.
Added method to select active walkable area in a room. Tested locally, it works.
Added methods to check which walkable_area is the active one.
Added buttons to the inspector to navigate between walkable area and its polygon instance. Have to be finished.
Added a flag (and related setget functions) to allow walking past a walkable area boundaries for a character.

FIX: I.add_item_as_active() wasn't working as expected

==== v1.8.2 ====
Mayor improvement in how interactions are registered in overlaping PopochiuClickables.
Now Popochiu stores an array of hovered PopochiuClickable to handle correctly what to show and which object is the clicked one.
Now CollisionPolygon2D nodes inside PopochiuClickables are called InteractionPolygon.
Created separated methods for calling `C.player_say()` and `C.character_say()` inside and outside an `E.run([])`.

FIX: `in_inventory` property changes after the PopochiuInventoryItem is added to the tree.
FIX: When characters talk using the "script" from the UI no longer quickly-appears-dissapears.

==== v1.8.1 ====
Added `link_to_item` property to PopochiuProp so the prop itself disables, enables when the item linked to it is added or discarded from the inventory.
Active inventory item deactivates with both left and right click.
Improvements in the flow of loading the room when running the project.
Created two stop methods in the AudioManager to stop a playing audio file inside a run and outside of one.
Props, Hotspots and Regions inside rooms are obtained by their group instead of checking the childrens of their corresponding container nodes.
IDialog now stores a reference to the current dialog and the last selected option.
IDialog now has a method for making the PC say the selected option: `D.say_selected()`.
Updated DialogTemplate.gd to include an example of calling `D.say_selected()`.
GraphicInterface and TransitionLayer folders are no longer moved to res://popochiu during Setup.
Updated Prop and Hotspot templates.
Improved how C.face_clicked() works. Now it calls the PopochiuCharacter.face_clicked() method.
When creating props with no interaction, the _NoInteraction folder is created inside the room's Props folder to place there all the assets for the non-interactable props.
AudioManager play methods now have tow options: one for when they run inside an E.run([]) and one for when they run out of a E.run([]).

FIX: Inspector plugin was throwing an error when editing a PopochiuDialog resource.
FIX: Setting a custom texture for Cursor can ignore blocking.
FIX: Clicking a row in the Room tab only selects the node in the Scene tree. It no longer tries to open its script file.
FIX: AudioCue by default uses the Master bus.
FIX: Removing characters and calling PopochiuRoom.on_room_exited no longer happens on the engine's first run.

==== v1.8.0 ====
* DisplayBox rendering improvements.

Now the node is a RichTextLabel.
Improvements in adjusting the size and position (centered to screen) of the node.

* Huge improvement in DialogText rendering.

No more additional Label needed (not at least fixed, it is created and deleted during runtime).
Now the continue icon renderds anchored to the size of its parent.

* GUI scales to match the game size.

This can be avoided with an option in PopochiuSettings.
Huge improvements for DisplayBox and DialogText rendering.
DisplayBox can now render BBcode.
Layout update for "Restart" popup.
Setup popup now shows a message advicing about the GUI scale.
Pixel game type set as default type after install.
Cursor and TransitionLayer also scale with the GUI.
Selecting a PopochiuCharacter node inside a PopochiuRoom now allows to edit the "position" property.

=== v1.7.2 ====
FIX: Plugin was not working because cyclic dependencies

==== v1.7.1 ====
PopochiuCharacter nodes can't be edited while selected inside a PopochiuRoom.
Added Warning message so players understand why PopochiuCharacter is not editable while in a PopochiuRoom.
Added two buttons to canvas editor menu for selecting a Prop or Hotspot baseline and walk to helpers.
When a PopochiuProp is created with interaction, or a PopochiuHotspot or a PopochiuRegion are created, a CollisionPolygon2D
is automatically created.
FIX: Deleting a Prop with no script will no longer delete its room folder.
FIX: Deleting a Prop with no interaction couldn't be deleted without restarting Godot.