IAudio and A scripts created to close #21
Now the AudioManager is added to the tree during runtime.
IAudio has the public methods that were in AudioManager and that were used to play AudioCues the old way.
Now devs can use A. to see a list with autocompletion for the AudioCues in the project.

FIX: Showing an inline dialog while a branching dialog is running no longer make the options of the second one to show while the logic of the first one is running.
FIX: Showing an inline dialog alone, or after an E.run([]) no longer blocks the graphic interface.