fix: solves issues related with AudioCues and stores character position in rooms
Fixes #23

`characters_cfg` no longer exists in PopochiuRoom. Now its state script has a `characters: Dictionary` property that stores the state of characters in the room.
At the moment, only `position` and `_looking_direction` are being stored.

FIX: #23 - Save/Load working

Saving and loading the game takes into account the data stored related to characters in the room.

Fix #31

Now rows created in TabAudio are sorted alphabetically when adding new audio files to the project.
AudioCues in PopochiuData.cfg are now sorted by file name.

Fixes #44

Added an EditorInspectorPlugin to handle the new PopochiuAudioCue class to show the list of buses currently configured for the project instead of the hard-coded ones.
The list updates each time the field is clicked in the Inspector, so adding a bus while an AudioCue is selected will update the list of buses immediately.
Assigned MxTwoPopochius to the new Music bus, and SfxToyCar to the Effects bus.

Now edition message in Room's characters is a button

Clicking it opens the scene of the character for edition.
EditorInspectorPlugins are now loaded different to allow the addition of new scripts easely.

Added fields to filter objects in Popochiu tabs

At the moment it only filters by the name seen in each row. A future improvement could be to take more properties into account for filtering.
Moved Inspector folder to Editor folder to match the upcoming project structure comming from feature/8-aseprite-importer.
PopochiuUtils is no longer loaded as an autoload. Now it is a global class with static methods.

WIP for improving the workflow when adding/removing characters in a room

Characters can be added/removed to room form Room tab

This will allow just to add or remove characters in a room, not to create or delete the PopochiuCharacter itself.
Updated how the delete ConfirmationDialog in PopochiuDock is resized.

* Rebased from #8 and #18 changes in main

Updated plugin version in plugin.cfg file.
Set PopodhiuDock min width to 392.

* Fix for importer defaults in Pixel project

CharacterTemplate.gd now calls `.play_xxx()` methods by default.
Adding a character to a room places in the center of it.
Removed custom fonts from Popochiu popups.
Filtering in Popochiu tabs now takes into account group titles.

* PR review changes

Reviewed-by: @stickgrinder 
Refs: #23 #31 #44