fix: fixes #58 #56 #17
Updated project rooms towards creating the game that looks like the one in Popochiu 1.
Fix #17: The method that looks for Popochiu objects is now called on start with a call_deferred.
Fix #56: This was because a change in the property name for BaseButton from `pressed` to `button_pressed`.
Fix #58: Now the prop that is added to the room's Props node is an instance of the created .tscn file.