Set agent_radius to 0 for NavigationPolygon in PopochiuWalkableArea
Godot by default sets it to 10, which makes the polygon to be not fully
used.
Listen EditorUndoRedoManager.history_changed to bake navigation polygons
in all PopochiuWalkableAreas so devs see the changes live without
needing to click the Bake NavigationPolygon button.
Update .gitignore commenting the lines related to the addons folder.