Store polygon for interaction in @export var
This will allow to keep the data of the polygon in the .tscn file of the
room so the InteractionPolygon.polygon data of each prop is overwritten
by the room both in runtime and while editing the room.

NOTE: The polygon of the InteractionPolygon in each prop is not saved in
the prop's scene file yet.