Fix for 170 and 166 (#172)
* Can cancel interactions when walking to clickable

This was achieved by checking the final position of the character once it stops moving against the target movement position of the clicked object.

[upd] Improved the way PopochiuClickable reacts to GUI blocking.
[upd] Improved the way the Cursor reacts to GUI blocking.

* Double check for cancelling `walk_to_clicked`

If the position check passes, then the `script_name` of the object that triggered the action is compared to the one of `E.clicked`.

* Reduce movement cancellation to just one comparisson

This is the one that checks if the clicked object is still active or if it didn't change to a new one (i.e. another PopochiuClickable which registered a click).