Add Popochiu settings sections to ProjectSettings
Parting from the progress made by @drbloop , now new sections for
PopochiuSettings are added to ProjectSettings when the Editor starts.
Next thing will be making sure values are properly saved and loaded, and
remove the GUI and TransitionLayer path dependency, which will be only
in the `popochiu_data.cfg` file.