Fix #219
Set `PopochiuRoom.state.visited_first_time` to `true` after the
`_on_room_transition_finished()` is executed.

- **fix** InventoryGrid slots were being deleted in the process of
adding items in the ProjectSettings.items_on_start array. This was
causing an issue in the 9 Verb and Sierra templates.
- **upd** Settings popup in 9 Verb and Sierra templates now close
automatically when a slot is selected for saving or loading.
- **upd** Slots in InventoryGrid are no longer deleted when removing an
item. Instead, only its name changes.
- **fix** Improve fix to avoid Godot printing a Warning related with
setting the size of dialog_text and hover_text during a `_ready()`
execution.