Refactor popochiu_object_row and remove dependencies to popochiu_dock (#236)
* Refactor scripts and classes in the Editor side

Remove main dock dependency in PopochiuObjectRow and PopochiuAudioRow

Create script that will serve as a signal bus to communicate components
in the Editor side of Popochiu

The DeleteConfirmation popup is created in runtime by PopochiuObjectRow
and PopochiuAudioRow when needed.

Improve use of DeleteConfirmation dialog

- **fix** issue when assigning the PC from popochiu_data.cfg.
- **fix** issue when clearing the inventory.
- **fix** deleting object rooms was not saving the .tscn file of the
room.

Separate the script of popochiu_object_row into several

Now popochiu_audio_row inherits from popochiu_row.

- **fea** Create popochiu_object_row specific scripts for rooms,
character, inventory item, and room object (props, hotspots, walkable
areas, regions and markers).
- **fea** Create Project Settings section for Popochiu audio. Here devs
we'll be able to define the prefix for audio cues.
- **upd** Remove main_dock dependency in TabAudio.
- **upd** Create folders for TabAudio and TabRoom.

Fix issue with character rows in Room tab

Also makes the Play button in popochiu_room_row visible.

- **upd** Update some calls to String.find() by `"TheStringToLookFor" in
String`.
- Audio tab now takes into account the prefixes in the audio section of
the Project Settings to clasify new audio cues.

Remove main_dock dependency in factories and creation popups

Remove non used functions

Remove `main_dock` dependency in Aseprite importer scripts

Separate the Main tab from the Popochiu dock. Now it has its own scene
and script.

Remove preload of PopochiuResources in Constants var

Now scripts access PopochiuResources directly.

- **upd** Minor flow improvement when removing main objects in
popochiu_object_row.

Add wait for popochiu_dock.ready signal

This makes a better flow for the plugin start, and avoid the use of
timers to assure everything is ready to show project data in the dock.

- **upd** Now the Setup popup is shown by calling a method in
PopochiuEditorHelper.
- **upd** The Audio tab connects directly to the signal needed to update
the list of audio files and cues, instead of the popochiu_plugin script
doing that.

Fix in Creation popup for room objects

- **fix** Fix Tab audio when deleting files from FileSystem.
- **upd** Change inheritance of Loading popup to AcceptDialog.
- **upd** Now Popochiu Editor popups are added to the dock, not to the
EditorInterface.
- **upd** Deleting a Popochiu Audio Cue now also deletes the audio file
if the checkbox is marked.

Remove popochiu_types_helper script

Now those functions are in PopochiuEditorHelper.

- **upd** Refactor Audio tab to separate the search for deleted audio
cues from the assignation of audio cues to their corresponding groups.
The deletion is a bit cleaner now.
- **fea** In the Popochiu/Audio section of the Project Settings devs can
specify the character that will be used to define the prefixes to
categorize audio files. Additionally, devs can now define different
prefixes for each category separating them by comma.

Progress on making popups work on Linux

Create scenes to put the content of some popups

Now instead of having a scene that inherits from ConfirmationDialog, the content of each dialog has its own scene, and the dialog is created by Popochiu when needed, putting the scene of said popup inside the created node. This improves the behavior of windows when shown.

Refactor for more object creation dialogs

Now characters, inventory items, dialogs, hotspots and markers are
created using the new dialogs.

- **upd** Improve the behavior of the creation dialog when its content
changes.

Refactor creation dialogs to remove repeated code

This makes each script shorter and specific.

Show error message when trying to create an existing object

When the path of the object to create is already in the FileSystem, the
Create Object Confirmation shows an error message and disables the OK
button of the dialog.

Update Loading and Setup dialogs to the new flow

Now the scenes for those popups only define the content of the popup, and it is added to the popup when it is created.

Fix Setup popup dissappearing on first GUI template copy process

Update folder and file names.

Apply suggestions from code review

Co-authored-by: Paolo Pustorino <stickgrinder@gmail.com>

Progress in PR review

Simplify the logic of playing/pausing audio rows.

Improve code based on PR review

Rename `show_add_to_core()` to `show_as_not_in_core()`

Remove the option to create the state script since this is no longer a
problem since version 1.9 or 1.10.

Refactor how is_main and is_pc are set.

This apply for popochiu_room_row and popochiu_character_row
correspondingly

- **upd** Remove type looking in popochiu_group by creating a public
function in popochiu_row that calls a virtual function so each row knows
if it should or not do something with clearing its tag icon.

Remove unnecessary editor validation in Audio tab

Fix an issue that was causing the PC not to be set properly when it was
selected in Popochiu's dock.

* Update to match defaults pattern of editor_config.gd