SimpleClick GUI improvements (#379)
* refs #373: Create bar and settings popup for the SimpleClickGUI.

These components will replace the old InventoryBar and SettingsBar as
part of the update for the SimpleClickGUI as proposed be Papernoise.

* refs #373: Make buttons and other components usable.

Players can now open the save, load, history and quit popups. They can
also change volumes, the text continue mode and the text speed.

* refs #373: Update icons and add setup options to Sound Volumes component.

Now devs can set if the Sound Volumens component should show the Master
bus only and if it should show the label with the name for each bus.

- **upd** Add custom icons for radio buttons in PopupMenu for the
  base_gui_theme.tres.

* refs #373: Create DialogueAdvancement component with better feedback.

This component uses a CheckButton with two labels so players can have
a better understanding of the current dialogue advancement mode.

* refs #373: Remove non used components and popups in SimpleClick GUI.

This also updates the behavior of the cursor when hovering the settings
button (to be improved).

* refs #373: Progress on migration script to update the SimpleClick GUI.

The script is executed only if the project is using the SimpleClick
template, and it removes the InventoryBar, SettingsBar,
TextSettingsPopup and SoundSettingsPopup components.

* refs #373: Update migration script to add new components and update the SaveAndLoadPopup.

The new Simple Click GUI uses the SimpleClickBar and the
SimpleClickPopup components. Instead of adding a new migration step, the
SaveAndLoadPopup is also removed in the first step and then added in the
second one.

- **upd** Remove the `@onready` vars in the Simple Click GUI to access
  the SaveAndLoadPopup, HistoryPopup, and QuitPopup since it is no
  longer used to open those popups.

* refs #373: Update migration to avoid removing the SaveAndLoadPopup

Instead of this, the popup's main script is updated with the content
of the one in the addons folder.

* refs #373: Remove commented code and improve saving

* refs #373: Improve and homogenize icon sizing when items enter inventory.

- **fix**: Assign unique names to the **Up** and **Down** icons in the
  base **InventoryGrid** component.

* refs #373: Improve property value assignation based on review suggestion.

Co-authored-by: Paolo Pustorino <stickgrinder@gmail.com>

* refs #373: Make improvements suggested by @stickgrinder on its PR review

- **upd** Reassign the texture used by the **AtlasTexture** of all the
  icons in the **SimpleClickSettingsPopup** to use the
  `simple_click_settings_popup_icons.png` instead of a
  CompressedTexture2D.

* refs #373: Update SimpleClickSettingsPopup icons based on PR feedback.

---------

Co-authored-by: Paolo Pustorino <stickgrinder@gmail.com>