Fixes #438 - Godot 4.4.1 crashing after plugin installation in Windows (#440)
* refs #438: Don't show setup popup on start.

Looks like something in the setup process is the cause of the engine
crashes.

- Fix UUID issues with some scenes.

* refs #438: Avoid saving opened scene and object row when removing Room objects.

The saving was ocurring twice, and the second attempt ocurred after the
row was deleted from the tree, which caused a crash.

* refs #438: Prevent Setup dialog from being removed after used

When the popup is deleted something causes the engine to crash when a
scene is saved. As a temporary fix, the Setup dialog will not be removed
but hidden instead.

The engine still crashes if the Setup window is opened after enablind the
plugin.

* refs #438: Show Setup popup after installing the plugin

Looks like it is working with the new approach for the Setup popup
handling.

* Update addons/popochiu/editor/main_dock/popochiu_row/object_row/room_object_row/popochiu_room_object_row.gd

Co-authored-by: Paolo Pustorino <stickgrinder@gmail.com>

* refs #438: Update code based on PR review comments.

---------

Co-authored-by: Paolo Pustorino <stickgrinder@gmail.com>