Recognize frame_changed from AnimatedSprite2D
Fix anti_glide_animation doesn't work if you change a Character's
Sprite2D to an AnimatedSprite2D.

We only need to ensure the signal exists so we can connect to it, so
check that instead of the type. This is only done on _ready, so it
shouldn't be a perf problem.

The Popochiu recommended method is to use a Sprite2D + AnimationPlayer,
but this code is specifically looking for a signal so check what we need
instead of what provides it.