doc: Put movement note after interaction explanation
The character won't walk to the hotspot until we script it (you only get
the can't INTERACT message), so the current line is wrong (and stumped
me for a bit!). Change it to the actual behaviour and move the "under
the hood" info box until when we actually get the character moving for
the first time.