refs #468: Rephrase and add detail to tutorial docs
* doc: Rephrase some confusing phrasing; finish hanging sentence

Rephrasing to make the intent more clear.

* doc: Minor cleanup in intro tutorial

Add missing marker for image 15. Add missing whitespace consistent with
other markers.

Fix malformed markdown for link and `code`.

Use same capitalization for bold "Walk To Point" as in-engine (this is
the only one that had capital "To" so not changing others).

* doc: Explain why editing polygon might not work

* doc: Put movement note after interaction explanation

The character won't walk to the hotspot until we script it (you only get
the can't INTERACT message), so the current line is wrong (and stumped
me for a bit!). Change it to the actual behaviour and move the "under
the hood" info box until when we actually get the character moving for
the first time.

* doc: Explain "Look At Point"

It wasn't clear to me that "Look At Point" means "this is where the
character looks" as opposed to "this is where we walk when trying to
look".

* doc: Mention how to change the main scene

I missed the checkbox on my first room and didn't find any mention in
the docs.